requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
Pillow==9.1.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
requests==2.28.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
